requirements.txt
asgiref==3.2.10
astroid==2.4.2
cffi==1.14.1
cryptography==3.0
Django==3.0.8
django-rest-knox==4.1.0
djangorestframework==3.11.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.5.3
pylint-django==2.2.0
pylint-plugin-utils==0.6
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
wrapt==1.12.1
